Pequenas correções
* Adicionado novas informações no /system help
* Modificados os comando /system backup e /system host avancado para rodarem em Async
* Adicionado novas informações no ajuda.yml
* Alterado o formato do nome do arquivo de backup do comando /system backup
* Outras pequenas alterações

+++ src/rush/addons/legendchat/McTopTag.java
@@ -38,7 +38,7 @@ public class McTopTag implements Listener {
 					playerTopOne = ((PlayerStat) tops.get(0)).name;
 				}
 			}
-		}.runTaskTimerAsynchronously(Main.get(), 60L, (long) Settings.mcTopTag_Tempo_De_Checagem * 20L);
+		}.runTaskTimerAsynchronously(Main.get(), 60L, Settings.mcTopTag_Tempo_De_Checagem * 20L);
 	}
 
 }
